backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
rembg==2.0.72
onnxruntime>=1.20.0
Pillow>=12.1.0
aiofiles==24.1.0
